Split out First tests to make things more managable
#68
